+++ infra/audioplayer/engine-ffmpeg/src/test/kotlin/app/campfire/audioplayer/engine/ffmpeg/FilterChainTest.kt
@@ -1,7 +1,7 @@
 // Copyright 2026, Drew Heavner and the Campfire project contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-package app.campfire.audioplayer.impl.engine.ffmpeg
+package app.campfire.audioplayer.engine.ffmpeg
 
 import assertk.assertThat
 import assertk.assertions.contains